+++ mystash-service/src/main/java/org/ranjangeorge/mystash/service/impl/LedgerEntry.java
@@ -4,7 +4,6 @@ import org.jetbrains.annotations.NotNull;
 import org.ranjangeorge.mystash.service.api.data.CreditOrDebit;
 
 import javax.persistence.*;
-import java.math.BigDecimal;
 import java.util.Date;
 import java.util.UUID;
 
@@ -21,7 +20,7 @@ public class LedgerEntry {
 
     private CreditOrDebit creditOrDebit;
 
-    private BigDecimal amount;
+    private long amount;
 
     private String description;
 
@@ -31,8 +30,9 @@ public class LedgerEntry {
     }
 
     public LedgerEntry(
-            Stash stash, @NotNull final CreditOrDebit creditOrDebit,
-            final BigDecimal amount,
+            Stash stash,
+            @NotNull final CreditOrDebit creditOrDebit,
+            final long amount,
             @NotNull final String description,
             @NotNull final Date date) {
 
@@ -57,7 +57,7 @@ public class LedgerEntry {
         return CreditOrDebit.DEBIT.equals(creditOrDebit);
     }
 
-    public BigDecimal getAmount() {
+    public long getAmount() {
         return amount;
     }
