+++ mystash-service/src/test/java/org/ranjangeorge/mystash/service/stashadmin/CreateStashTest.java
@@ -7,7 +7,6 @@ import org.junit.Test;
 import org.ranjangeorge.mystash.service.api.IStashAdminService;
 import org.ranjangeorge.mystash.service.impl.support.ServiceLookup;
 
-import java.math.BigDecimal;
 import java.util.Set;
 
 public class CreateStashTest {
@@ -34,7 +33,7 @@ public class CreateStashTest {
         stashAdminService.createNewStash(
                 "my-new-stash",
                 "george.ranjan@gmail.com",
-                BigDecimal.valueOf(100d));
+                100L);
         //
         Set<String> stashes = stashAdminService.listAllStashes();
         //
